Refactor vm memory endpoints to separate service part
